Add files via upload
In this project I applied boosting techniques (Ensemble, Random Rorest, Ada boosting, Gradient boosting, XBG and Ada Boosting), balancing with imbalanced learn, and LIME for feature analysis.
I am going to work with a dataset provided in Kaggle by "Banco Santander", where they describe the dataset as:
"You are provided with an anonymized dataset containing numeric feature variables, the binary target column, and a string ID_code column.
The task is to predict the value of target column in the test set."
Was uploaded as a competition on the site 4 years ago: "In this challenge, we invite Kagglers to help us identify which customers will make a specific transaction in the future, irrespective of the amount of money transacted. The data provided for this competition has the same structure as the real data we have available to solve this problem".